永豐/linebot/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
python-dotenv==1.0.1
line-bot-sdk==3.11.0
google-generativeai==0.5.0